Move implementation details from tags.cpp to tags_int.cpp
