+++ src/CMakeLists.txt
@@ -113,6 +113,7 @@ set( LIBEXIV2_PRIVATE_SRC
     samsungmn_int.cpp
     sigmamn_int.cpp
     sonymn_int.cpp
+    tags_int.cpp
     tiffcomposite_int.cpp
     tiffimage_int.cpp
     tiffvisitor_int.cpp
